+++ src/components/Loader/index.jsx
@@ -1,4 +1,5 @@
 function Loader() {
+  // Todo: center loader
   const circleCommonClasses = "h-2.5 w-2.5 bg-current rounded-full";
 
   return (
